requirements-dev.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.3.0
black==20.8b1
click==7.1.2
coverage==5.3.1
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.8
paho-mqtt==1.5.1
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
regex==2020.11.13
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.4.0
